revert babel back due to react-dialog requirements. Server-side test fixes to accomodate refactoring
