+++ frontend/TODO
@@ -1,0 +1,1 @@
+replace react-dialog due to its reliance on outdated dependencies
