+++ test/IngestProxyQueueSpec.scala
@@ -3,7 +3,7 @@ import java.time.ZonedDateTime
 import akka.actor.Props
 import akka.testkit.TestProbe
 import com.theguardian.multimedia.archivehunter.common._
-import com.theguardian.multimedia.archivehunter.common.clientManagers.{DynamoClientManager, S3ClientManager, SQSClientManager}
+import com.theguardian.multimedia.archivehunter.common.clientManagers.{DynamoClientManager, ESClientManager, S3ClientManager, SQSClientManager}
 import com.theguardian.multimedia.archivehunter.common.cmn_models.{IngestMessage, ScanTargetDAO}
 import org.specs2.mock.Mockito
 import org.specs2.mutable.Specification
@@ -38,13 +38,15 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockS3ClientMgr = mock[S3ClientManager]
       val mockDDBlientMgr = mock[DynamoClientManager]
       val mockDDBClient = mock[DynamoClient]
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
       mockDDBlientMgr.getNewAlpakkaDynamoClient(any)(any, any) returns mockDDBClient
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -72,12 +74,14 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockDDBlientMgr = mock[DynamoClientManager]
       val mockDDBClient = mock[DynamoClient]
       mockDDBlientMgr.getNewAlpakkaDynamoClient(any)(any, any) returns mockDDBClient
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -105,11 +109,13 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockDDBClient = mock[DynamoClient]
       mockDDBlientMgr.getNewAlpakkaDynamoClient(any)(any, any) returns mockDDBClient
       val mockEtsProxyActor = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -147,9 +153,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -181,9 +189,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr,mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -215,9 +225,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -251,9 +263,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -283,9 +297,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -318,9 +334,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -356,9 +374,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -389,9 +409,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -444,9 +466,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
 
@@ -472,9 +496,11 @@ class IngestProxyQueueSpec extends Specification with Mockito with ZonedDateTime
       val mockEtsProxyActor = TestProbe()
       implicit val mockScanTargetDAO = mock[ScanTargetDAO]
       val mockedSelf = TestProbe()
+      val mockESClientMgr = mock[ESClientManager]
+      implicit val mockIndexer = mock[Indexer]
 
       val toTest = system.actorOf(Props(new IngestProxyQueue(testConfig, system, mockSqsClientMgr, mockProxyGenerators, mockS3ClientMgr,
-        mockDDBlientMgr) {
+        mockDDBlientMgr, mockESClientMgr) {
         override protected val ownRef = mockedSelf.ref
       }))
