+++ test/ProxyFrameworkQueueSpec.scala
@@ -312,133 +312,6 @@ class ProxyFrameworkQueueSpec extends Specification with Mockito {
     }
   }
 
-  "ProxyFrameworkQueue.setProxiedWithRetry" should {
-    "retry the index operation if a version conflict error occurs" in {
-      import scala.concurrent.ExecutionContext.Implicits.global
-      val fakeEntry = ArchiveEntry(
-        "some-item-id",
-        "some-bucket",
-        "path/to/item",
-        None,
-        None,
-        1234L,
-        ZonedDateTime.now(),
-        "some-etag",
-        MimeType("video","mp4"),
-        false,
-        StorageClass.STANDARD,
-        Seq(),
-        beenDeleted = false,
-        mediaMetadata = None
-      )
-
-      val initialError = RequestFailure(409,None,Map(),ElasticError("version_conflict_engine_exception","something went splat",None,None,None,Seq()))
-
-      val mockIndexer = mock[Indexer]
-      mockIndexer.indexSingleItem(any,any,any)(any) returns Future(Left(initialError)) thenReturns Future(Left(initialError)) thenReturns Future(Right("someid"))
-      mockIndexer.getById(any)(any) returns Future(fakeEntry)
-
-      val mockLogger = Logger("test")
-
-      val mockEsClient = mock[HttpClient]
-
-      class TestClass extends ProxyFrameworkQueueFunctions {
-        override protected val indexer = mockIndexer
-        override protected val logger = mockLogger
-
-        override protected val esClient = mock[HttpClient]
-      }
-
-      val toTest = new TestClass
-      val result = Await.result(toTest.setProxiedWithRetry("some-source-id"), 30 seconds)
-      result must beRight("someid")
-      there were three(mockIndexer).indexSingleItem(any, any, any)(any)
-      there were three(mockIndexer).getById(any)(any)
-    }
-
-    "just return the value if the operation succeeds" in {
-      import scala.concurrent.ExecutionContext.Implicits.global
-      val fakeEntry = ArchiveEntry(
-        "some-item-id",
-        "some-bucket",
-        "path/to/item",
-        None,
-        None,
-        1234L,
-        ZonedDateTime.now(),
-        "some-etag",
-        MimeType("video","mp4"),
-        false,
-        StorageClass.STANDARD,
-        Seq(),
-        beenDeleted = false,
-        mediaMetadata = None
-      )
-
-      val mockIndexer = mock[Indexer]
-      mockIndexer.indexSingleItem(any,any,any)(any) returns Future(Right("someid"))
-      mockIndexer.getById(any)(any) returns Future(fakeEntry)
-
-      val mockLogger = Logger("test")
-
-      class TestClass extends ProxyFrameworkQueueFunctions {
-        override protected val indexer = mockIndexer
-        override protected val logger = mockLogger
-
-        override protected val esClient = mock[HttpClient]
-      }
-
-      val toTest = new TestClass
-      val result = Await.result(toTest.setProxiedWithRetry("some-source-id"), 30 seconds)
-      result must beRight("someid")
-      there was one(mockIndexer).indexSingleItem(any, any, any)(any)
-      there was one(mockIndexer).getById(any)(any)
-    }
-
-    "pass back an error that is not a conflict error" in {
-      import scala.concurrent.ExecutionContext.Implicits.global
-      val fakeEntry = ArchiveEntry(
-        "some-item-id",
-        "some-bucket",
-        "path/to/item",
-        None,
-        None,
-        1234L,
-        ZonedDateTime.now(),
-        "some-etag",
-        MimeType("video","mp4"),
-        false,
-        StorageClass.STANDARD,
-        Seq(),
-        beenDeleted = false,
-        mediaMetadata = None
-      )
-
-      val initialError = RequestFailure(500,None,Map(),ElasticError("some_other_error","something went splat",None,None,None,Seq()))
-
-      val mockIndexer = mock[Indexer]
-      mockIndexer.indexSingleItem(any,any,any)(any) returns Future(Left(initialError))
-      mockIndexer.getById(any)(any) returns Future(fakeEntry)
-
-      val mockLogger = Logger("test")
-
-      val mockEsClient = mock[HttpClient]
-
-      class TestClass extends ProxyFrameworkQueueFunctions {
-        override protected val indexer = mockIndexer
-        override protected val logger = mockLogger
-
-        override protected val esClient = mock[HttpClient]
-      }
-
-      val toTest = new TestClass
-      val result = Await.result(toTest.setProxiedWithRetry("some-source-id"), 30 seconds)
-      result must beLeft
-      there was one(mockIndexer).indexSingleItem(any, any, any)(any)
-      there was one(mockIndexer).getById(any)(any)
-    }
-  }
-
   "ProxyFrameworkQueue.convertMessageBody" should {
     "parse the message body and return a domain object, including the message timestamp" in new AkkaTestkitSpecs2Support {
       val msgContent =
